backend/backend/requirements.dev.txt
anyio
Django~=4.2.13
# celery~=5.4.0
# django-celery-beat
# django-celery-results
django-cors-headers
django-extensions
django-filter
django-js-asset
django-oauth-toolkit
django-polymorphic
django-redis
django-rest-passwordreset
django-rest-polymorphic
django-timezone-field
djangorestframework~=3.15.2
dj-rest-auth 
djangorestframework-simplejwt
django-allauth~=0.63.3
drf-nested-routers
drf-standardized-errors
oauthlib
pillow
psutil
psycopg
psycopg2-binary
PyJWT
pyparsing
python-dateutil
python-dotenv
PyYAML
redis
regex
requests
six
gunicorn
stripe~=10.0.0
django-parler
django-parler-rest~=2.2
django-debug-toolbar~=4.4.2
django-gmailapi-backend
django_grapesjs
django-notifications-hq~=1.8.3
django-jazzmin
django-environ
jsonfield
django-log-viewer
django-soft-delete
django-restful-admin



numpy
pandas
tqdm
Pillow
# scipy
tensorflow>=2.0.0
# opencv-python>=4.0.0
imageio>=2.0.0
imgaug>=0.4.0
# scikit-learn>=0.22.0
# matplotlib>=3.0.0
pyheif==0.5.1
clip-by-openai
# torch==1.7.1
whatimage==0.0.3
pyaml
